// src/server/types.gen.ts
export type ModelsAnalyticsOverview = {
    bound_players?: number;
    daily_active?: number;
    online_rate?: number;
    online_servers?: number;
    storage_used?: number;
    total_players?: number;
    total_servers?: number;
    total_uploads?: number;
};